+++ tests/array_test.py
@@ -30,7 +30,6 @@ from jax._src import op_shardings
 from jax._src import test_util as jtu
 from jax._src import xla_bridge as xb
 from jax._src.lib import xla_client as xc
-from jax._src.lib import xla_extension_version
 from jax._src.util import safe_zip
 from jax._src.sharding_impls import (_op_sharding_to_pos_sharding,
                                      pmap_sharding_devices_indices_map)
@@ -790,7 +789,6 @@ class JaxArrayTest(jtu.JaxTestCase):
     self.assertTrue(arr.is_deleted())
 
   @parameterized.product(dtype=jtu.dtypes.all + jtu.dtypes.custom_floats)
-  @unittest.skipIf(xla_extension_version < 208, "Test requires jaxlib > 0.4.19")
   def test_shards_have_correct_dtype(self, dtype):
     x = jnp.ones((), dtype=dtype)
     for shard in x.addressable_shards: